Rename libcfpub to libcompat (Redmine #2052)
libcompat is a well-known name, so stick to it.
