+++ ext/Makefile.am
@@ -1,4 +1,4 @@
 sbin_PROGRAMS = rpmvercmp
 
-LDADD = ../pub/libcfpub.la
+LDADD = ../libcompat/libcompat.la
