+++ libcompat/drand48.c
