+++ libcompat/getloadavg.c
