+++ libcompat/getopt.c
