+++ libcompat/getopt.h
