+++ libcompat/getopt1.c
