+++ libcompat/memdup.c
