+++ libcompat/mkdtemp.c
