+++ libcompat/pthread_attr_setstacksize.c
