+++ libcompat/setlinebuf.c
