+++ libcompat/snprintf.c
