+++ libcompat/srand48.c
