+++ libcompat/strcasecmp.c
