+++ libcompat/strcasestr.c
