+++ libcompat/strdup.c
