+++ libcompat/strerror.c
