+++ libcompat/strlcat.c
