+++ libcompat/strlcpy.c
