+++ libcompat/strncasecmp.c
