+++ libcompat/strndup.c
