+++ libcompat/strnlen.c
