+++ libcompat/strrstr.c
