+++ libcompat/strsep.c
