+++ libcompat/strsignal.c
