+++ libcompat/strstr.c
