+++ libpromises/patches.c
@@ -28,7 +28,7 @@
   given platform These are conditionally compiled, pending extensions or
   developments in the OS concerned.
 
-  FIXME: move to the pub/ directory or to the apropriate source file.
+  FIXME: move to the libcompat/ directory or to the apropriate source file.
 */
 
 #include "cf3.defs.h"
